refactor(bp): Move common functions in BP folder to shared
