+++ etl/steps/data/garden/bp/2022-07-14/energy_mix.py
@@ -2,24 +2,19 @@
 
 """
 
-from copy import deepcopy
-from pathlib import Path
-from typing import cast
-
 import numpy as np
 import pandas as pd
-from owid.datautils import geo
 
 from etl.paths import DATA_DIR
 from owid import catalog
+from shared import CURRENT_DIR, add_population
 
-NAMESPACE = "bp"
-VERSION = 2022
-# Original BP's Statistical Review Dataset name in the owid catalog (without the institution and year).
-DATASET_CATALOG_NAME = "statistical_review_of_world_energy"
-NAMESPACE_IN_CATALOG = "bp_statreview"
+# Namespace, dataset short name, and version of the garden dataset of the BP statistical review.
+STAT_REVIEW_NAMESPACE = "bp"
+STAT_REVIEW_SHORT_NAME = "bp_statistical_review"
+STAT_REVIEW_VERSION = "2022-07-14"
 # Path to metadata file.
-METADATA_FILE_PATH = Path(__file__).parent / "energy_mix.meta.yml"
+METADATA_FILE_PATH = CURRENT_DIR / "energy_mix.meta.yml"
 
 # Conversion factors.
 # Terawatt-hours to kilowatt-hours.
@@ -42,225 +37,6 @@ DIRECT_AND_EQUIVALENT_ENERGY = [
 ]
 ALL_SOURCES = sorted(ONLY_DIRECT_ENERGY + DIRECT_AND_EQUIVALENT_ENERGY)
 
-REGIONS_TO_ADD = {
-    "North America": {
-        "area_code": "OWID_NAM",
-    },
-    "South America": {
-        "area_code": "OWID_SAM",
-    },
-    "Europe": {
-        "area_code": "OWID_EUR",
-    },
-    "European Union (27)": {
-        "area_code": "OWID_EU27",
-    },
-    "Africa": {
-        "area_code": "OWID_AFR",
-    },
-    "Asia": {
-        "area_code": "OWID_ASI",
-    },
-    "Oceania": {
-        "area_code": "OWID_OCE",
-    },
-    "Low-income countries": {
-        "area_code": "OWID_LIC",
-    },
-    "Upper-middle-income countries": {
-        "area_code": "OWID_UMC",
-    },
-    "Lower-middle-income countries": {
-        "area_code": "OWID_LMC",
-    },
-    "High-income countries": {
-        "area_code": "OWID_HIC",
-    },
-}
-
-# When creating region aggregates, decide how to distribute historical regions.
-# The following decisions are based on the current location of the countries that succeeded the region, and their income
-# group. Continent and income group assigned corresponds to the continent and income group of the majority of the
-# population in the member countries.
-HISTORIC_TO_CURRENT_REGION = {
-    "USSR": {
-        "continent": "Europe",
-        "income_group": "Upper-middle-income countries",
-        "members": [
-            # Europe - High-income countries.
-            "Lithuania",
-            "Estonia",
-            "Latvia",
-            # Europe - Upper-middle-income countries.
-            "Moldova",
-            "Belarus",
-            "Russia",
-            # Europe - Lower-middle-income countries.
-            "Ukraine",
-            # Asia - Upper-middle-income countries.
-            "Georgia",
-            "Armenia",
-            "Azerbaijan",
-            "Turkmenistan",
-            "Kazakhstan",
-            # Asia - Lower-middle-income countries.
-            "Kyrgyzstan",
-            "Uzbekistan",
-            "Tajikistan",
-        ],
-    },
-}
-
-
-def load_population() -> pd.DataFrame:
-    """Load OWID population dataset, and add historical regions to it.
-
-    Returns
-    -------
-    population : pd.DataFrame
-        Population dataset.
-
-    """
-    # Load population dataset.
-    population = (
-        catalog.find("population", namespace="owid", dataset="key_indicators")
-        .load()
-        .reset_index()[["country", "year", "population"]]
-    )
-
-    # Add data for historical regions (if not in population) by adding the population of its current successors.
-    countries_with_population = population["country"].unique()
-    missing_countries = [
-        country
-        for country in HISTORIC_TO_CURRENT_REGION
-        if country not in countries_with_population
-    ]
-    for country in missing_countries:
-        members = HISTORIC_TO_CURRENT_REGION[country]["members"]
-        _population = (
-            population[population["country"].isin(members)]
-            .groupby("year")
-            .agg({"population": "sum", "country": "nunique"})
-            .reset_index()
-        )
-        # Select only years for which we have data for all member countries.
-        _population = _population[_population["country"] == len(members)].reset_index(
-            drop=True
-        )
-        _population["country"] = country
-        population = pd.concat(
-            [population, _population], ignore_index=True
-        ).reset_index(drop=True)
-
-    error = "Duplicate country-years found in population. Check if historical regions changed."
-    assert population[population.duplicated(subset=["country", "year"])].empty, error
-
-    return cast(pd.DataFrame, population)
-
-
-def load_income_groups() -> pd.DataFrame:
-    """Load dataset of income groups and add historical regions to it.
-
-    Returns
-    -------
-    income_groups : pd.DataFrame
-        Income groups data.
-
-    """
-    income_groups = (
-        catalog.find(
-            table="wb_income_group",
-            dataset="wb_income",
-            namespace="wb",
-            channels=["garden"],
-        )
-        .load()
-        .reset_index()
-    )
-    # Add historical regions to income groups.
-    for historic_region in HISTORIC_TO_CURRENT_REGION:
-        historic_region_income_group = HISTORIC_TO_CURRENT_REGION[historic_region][
-            "income_group"
-        ]
-        if historic_region not in income_groups["country"]:
-            historic_region_df = pd.DataFrame(
-                {
-                    "country": [historic_region],
-                    "income_group": [historic_region_income_group],
-                }
-            )
-            income_groups = pd.concat(
-                [income_groups, historic_region_df], ignore_index=True
-            )
-
-    return cast(pd.DataFrame, income_groups)
-
-
-def add_population(
-    df: pd.DataFrame,
-    country_col: str = "country",
-    year_col: str = "year",
-    population_col: str = "population",
-    warn_on_missing_countries: bool = True,
-    show_full_warning: bool = True,
-) -> pd.DataFrame:
-    """Add a column of OWID population to the countries in the data, including population of historical regions.
-
-    This function has been adapted from datautils.geo, because population currently does not include historic regions.
-    We include them in this function.
-
-    Parameters
-    ----------
-    df : pd.DataFrame
-        Data without a column for population (after harmonizing elements, items and country names).
-    country_col : str
-        Name of country column in data.
-    year_col : str
-        Name of year column in data.
-    population_col : str
-        Name for new population column in data.
-    warn_on_missing_countries : bool
-        True to warn if population is not found for any of the countries in the data.
-    show_full_warning : bool
-        True to show affected countries if the previous warning is raised.
-
-    Returns
-    -------
-    df_with_population : pd.DataFrame
-        Data after adding a column for population for all countries in the data.
-
-    """
-
-    # Load population dataset.
-    population = load_population().rename(
-        columns={
-            "country": country_col,
-            "year": year_col,
-            "population": population_col,
-        }
-    )[[country_col, year_col, population_col]]
-
-    # Check if there is any missing country.
-    missing_countries = set(df[country_col]) - set(population[country_col])
-    if len(missing_countries) > 0:
-        if warn_on_missing_countries:
-            geo.warn_on_list_of_entities(
-                list_of_entities=missing_countries,
-                warning_message=(
-                    f"{len(missing_countries)} countries not found in population"
-                    " dataset. They will remain in the dataset, but have nan"
-                    " population."
-                ),
-                show_list=show_full_warning,
-            )
-
-    # Add population to original dataframe.
-    df_with_population = pd.merge(
-        df, population, on=[country_col, year_col], how="left"
-    )
-
-    return df_with_population
-
 
 def get_bp_data(bp_table: catalog.Table) -> pd.DataFrame:
     """Extract a simple dataframe of BP statistical review data from the table in the dataset.
@@ -280,13 +56,13 @@ def get_bp_data(bp_table: catalog.Table) -> pd.DataFrame:
 
     # Convert table (snake case) column names to human readable names.
     bp_table = bp_table.rename(
-        columns={column: bp_table[column].metadata.title for column in bp_table.columns}
+        columns={column: bp_table[column].metadata.title for column in bp_table.columns if column != "country_code"}
     ).reset_index()
 
     # Rename human-readable columns (and select only the ones that will be used).
     columns = {
-        "entity_name": "Country",
-        "entity_code": "Country code",
+        "country": "Country",
+        "country_code": "Country code",
         "year": "Year",
         # Fossil fuel primary energy (in EJ).
         "Coal Consumption - EJ": "Coal (EJ)",
@@ -311,7 +87,8 @@ def get_bp_data(bp_table: catalog.Table) -> pd.DataFrame:
     }
 
     # Create a simple dataframe (without metadata and with a dummy index).
-    assert set(columns) < set(bp_table.columns), "Column names have changed in BP data."
+    assert set(columns) < set(bp_table.columns), "Column names have changed in BP data."    
+
     bp_data = (
         pd.DataFrame(bp_table)[list(columns)]
         .rename(errors="raise", columns=columns)
@@ -572,62 +349,6 @@ def calculate_primary_energy_annual_change(
     return primary_energy
 
 
-def add_region_aggregates(primary_energy: pd.DataFrame) -> pd.DataFrame:
-    """Add region aggregate for all regions.
-
-    Regions are defined above, in REGIONS_TO_ADD.
-
-    Parameters
-    ----------
-    primary_energy : pd.DataFrame
-        BP data.
-
-    Returns
-    -------
-    primary_energy : pd.DataFrame
-        BP data after adding regions.
-
-    """
-    primary_energy = primary_energy.copy()
-
-    income_groups = load_income_groups()
-    aggregates = {
-        column: "sum"
-        for column in primary_energy.columns
-        if column not in ["Country", "Year", "Country code"]
-    }
-
-    for region in REGIONS_TO_ADD:
-        countries_in_region = geo.list_countries_in_region(
-            region=region, income_groups=income_groups
-        )
-        # We do not impose a list of countries that must have data, because, for example, prior to 1985, there is
-        # no data for Russia, and therefore we would have no data for Europe.
-        # Also, for a similar reason, the fraction of nans allowed in the data is increased to avoid losing all
-        # data in Europe prior to 1985.
-        # Similarly, we do not impose any maximum fraction of nans when aggregating, because there is usually a
-        # small number of African countries informed. Either we accept that "Africa" is represented by a small number
-        # of countries (which changes for different variables) or we have no data for "Africa".
-        # We choose the former.
-        primary_energy = geo.add_region_aggregates(
-            df=primary_energy,
-            region=region,
-            country_col="Country",
-            year_col="Year",
-            aggregations=aggregates,
-            countries_in_region=countries_in_region,
-            countries_that_must_have_data=[],
-            frac_allowed_nans_per_year=None,
-            num_allowed_nans_per_year=None,
-        )
-        # Add country code for current region.
-        primary_energy.loc[
-            primary_energy["Country"] == region, "Country code"
-        ] = REGIONS_TO_ADD[region]["area_code"]
-
-    return primary_energy
-
-
 def add_per_capita_variables(primary_energy: pd.DataFrame) -> pd.DataFrame:
     """Add per-capita variables.
 
@@ -743,107 +464,20 @@ def prepare_output_table(primary_energy: pd.DataFrame) -> catalog.Table:
     return table
 
 
-def load_table_from_previous_dataset() -> catalog.Table:
-    """Load table from previous version of the dataset of the current step.
-
-    Returns
-    -------
-    table_old : catalog.Table
-        Table from the previous version of the current dataset to be created.
-
-    """
-    # Sort the paths of candidate datasets from newest to oldest.
-    dataset_paths = sorted((DATA_DIR / "garden" / NAMESPACE).glob("*/energy_mix"))[::-1]
-
-    # Load one by one each dataset and stop if the version is the one from previous year.
-    for dataset_path in dataset_paths:
-        dataset = catalog.Dataset(dataset_path)
-        if dataset.metadata.version == VERSION - 1:
-            break
-    # Extract the (expected only one) table in that dataset.
-    table_old = dataset[dataset.table_names[0]]
-
-    return table_old
-
-
-def fill_missing_values_with_previous_version(
-    table: catalog.Table, table_old: catalog.Table
-) -> catalog.Table:
-    """Fill missing values in current data with values from the previous version of the dataset.
-
-    Parameters
-    ----------
-    table : catalog.Table
-        Processed data from current dataset.
-    table_old : catalog.Table
-        Processed data from previous dataset.
-
-    Returns
-    -------
-    combined : catalog.Table
-        Combined table, with data from the current data, but after filling missing values with data from the previous
-        version of the dataset.
-
-    """
-    # For region aggregates, avoid filling nan with values from previous releases.
-    # The reason is that aggregates each year may include data from different countries.
-    # This is especially necessary in 2022 because regions had different definitions in 2021 (the ones by BP).
-    # Remove region aggregates from the old table.
-    table_old = table_old.drop(table_old.loc[list(REGIONS_TO_ADD)].index).copy()
-
-    # Combine the current output table with the table from the previous version the dataset.
-    combined = pd.merge(
-        table,
-        table_old.drop(columns="country_code"),
-        left_index=True,
-        right_index=True,
-        how="left",
-        suffixes=("", "_old"),
-    )
-    # List the common columns that can be filled with values from the previous version.
-    columns = [column for column in combined.columns if column.endswith("_old")]
-    # Fill missing values in the current table with values from the old table.
-    for column_old in columns:
-        column = column_old.replace("_old", "")
-        combined[column] = combined[column].fillna(combined[column_old])
-    # Remove columns from the old table.
-    combined = combined.drop(columns=columns)
-
-    # Transfer metadata from the table of the current dataset into the combined table.
-    combined.metadata = deepcopy(table.metadata)
-    for column in combined.columns:
-        combined[column].metadata = deepcopy(table[column].metadata)
-
-    # Sanity checks.
-    assert len(combined) == len(table)
-    assert set(combined.columns) == set(table.columns)
-
-    return combined
-
-
 def run(dest_dir: str) -> None:
     #
     # Load data.
     #
-    # Load table from latest BP dataset.
-    bp_table = catalog.find_one(
-        DATASET_CATALOG_NAME,
-        channels=["backport"],
-        namespace=f"{NAMESPACE_IN_CATALOG}@{VERSION}",
-    )
-
-    # Load previous version of the BP energy mix dataset, that will be used at the end to fill missing values in the
-    # current dataset.
-    table_old = load_table_from_previous_dataset()
+    # Load the latest BP statistical review.
+    bp_dataset_path = DATA_DIR / "garden" / STAT_REVIEW_NAMESPACE / STAT_REVIEW_VERSION / STAT_REVIEW_SHORT_NAME
+    bp_dataset = catalog.Dataset(bp_dataset_path)
+    bp_table = bp_dataset[bp_dataset.table_names[0]]
 
     #
     # Process data.
     #
     # Get a dataframe out of the BP table.
     primary_energy = get_bp_data(bp_table=bp_table)
-
-    # Add region aggregates.
-    primary_energy = add_region_aggregates(primary_energy=primary_energy)
 
     # Calculate direct and primary energy using the substitution method.
     primary_energy = calculate_direct_primary_energy(primary_energy=primary_energy)
@@ -861,11 +495,6 @@ def run(dest_dir: str) -> None:
     # Prepare output data in a convenient way.
     table = prepare_output_table(primary_energy)
 
-    # Fill missing values in current table with values from the previous dataset, when possible.
-    combined = fill_missing_values_with_previous_version(
-        table=table, table_old=table_old
-    )
-
     #
     # Save outputs.
     #
@@ -877,9 +506,9 @@ def run(dest_dir: str) -> None:
     dataset.save()
 
     # Add table to the dataset.
-    combined.metadata.title = dataset.metadata.title
-    combined.metadata.description = dataset.metadata.description
-    combined.metadata.dataset = dataset.metadata
-    combined.metadata.short_name = dataset.metadata.short_name
-    combined.metadata.primary_key = list(combined.index.names)
-    dataset.add(combined, repack=True)
+    table.metadata.title = dataset.metadata.title
+    table.metadata.description = dataset.metadata.description
+    table.metadata.dataset = dataset.metadata
+    table.metadata.short_name = dataset.metadata.short_name
+    table.metadata.primary_key = list(table.index.names)
+    dataset.add(table, repack=True)
